#160: add option --name or -n for running specific tests by giving a test name
By using --name or -n we can specify the test method we want to run. We can
also give a pattern for test names that we want to run. To give a pattern,
just enclose the string between slashes (/). When the given test name or
pattern does not match any test, valid test names will be printed and the
process will abort.