Make socket location configurable
On vagrant using the rails root as the socket location is quite a hassle
since by default synced folders are shared so that creating a socket
inside is not possible. By making the location configurable via ZEUSSOCK
environment variable, such usage get way easier.

I moved the logic for the location to unixsocket to have it in a
location already referenced by both using modules, and not duplicating
it as it was before by setting the const string.