Log zeus error to stderr
When zeus has an error it is quite inconvinient to have it written to stdout as
this easily hides problems. To resolve this we allow `shinylog` to log to
`stderr` for strings. As those are user facing error we want to hide the
implementation detail of go so don't log error locations as they are most likely
not helpful to the end user.