Monitor directories on OSx Instead of files
OSX is much more limited in the amount of files to keep open compared to other
unixes since Sierra. So instead of monitoring files we monitor directories and
assume a change in the directory warents a restart.

This is less than ideal but should provide a quick fix for the current problem
on Sierra erroring on large applications.